--- internal/app/api/domain/repository/helper/transaction.go
@@ -1,8 +1,0 @@
-//go:generate mockgen -source=$GOFILE -package=$GOPACKAGE -destination=../../../../../../test/mock/domain/repository/$GOPACKAGE/$GOFILE
-package helper
-
-import "context"
-
-type TransactionObject interface {
-	Transaction(context.Context, func(context.Context) error) error
-}
