+++ test/mock/domain/repository/pkg/transaction/transaction.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=transaction.go -package=helper -destination=../../../../../../test/mock/domain/repository/helper/transaction.go
+//	mockgen -source=transaction.go -package=transaction -destination=../../../../../../../test/mock/domain/repository/pkg/transaction/transaction.go
 //
 
-// Package helper is a generated GoMock package.
-package helper
+// Package transaction is a generated GoMock package.
+package transaction
 
 import (
 	context "context"
